add jsdoc and move capability to before addon
