+++ lib/capabilities/kro-capability.ts
@@ -1,17 +1,33 @@
-import { Capability, CapabilityProps, CapabilityType } from "./capability";
+import { CapabilityType } from "../spi";
+import { Capability, CapabilityProps } from "./capability";
 
+/**
+ * Properties for KroCapability, excluding namespace, type, and identityCenterArn
+ * which are not applicable for KRO capabilities.
+ */
 export type KroCapabilityProps = Omit<CapabilityProps, "namespace" | "type" | "identityCenterArn">;
 
-
+/**
+ * KRO (Kubernetes Resource Orchestrator) capability for EKS clusters.
+ * Enables advanced resource orchestration and management within Kubernetes.
+ * });
+ * ```
+ */
 export class KroCapability extends Capability {
   readonly DEFAULT_POLICY_NAME = undefined;
+  
+  /** Default configuration for KRO capabilities */
   static readonly defaultProps: KroCapabilityProps = {
     useDefaultPolicy: true,
     capabilityName: "blueprints-kro-cap"
   };
 
+  /**
+   * Creates a new KRO capability instance.
+   * 
+   * @param options - Configuration options for the KRO capability
+   */
   constructor(readonly options: KroCapabilityProps){
     super({...KroCapability.defaultProps, ...options, type: CapabilityType.KRO});
   }
-
 }
