+++ lib/spi/capability-contracts.ts
@@ -1,6 +1,20 @@
 import { CfnCapability } from "aws-cdk-lib/aws-eks";
-import { ClusterInfo } from "./types";
+import { CapabilityType, ClusterInfo } from "./types";
+import { Construct } from "constructs";
 
+/**
+ * ClusterCapability is the interface to which all EKS Capabilities will conform.
+ */
 export declare interface ClusterCapability {
-  create(clusterInfo: ClusterInfo): CfnCapability
+  /**
+   * CapabilityType
+   */
+  type:  CapabilityType
+
+  /**
+   * Creates the capability on the cluster
+   * @param clusterInfo
+   * @returns cdk.Construct
+   */
+  create(clusterInfo: ClusterInfo): Construct
 }
